requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
blinker==1.7.0
bt==1.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
et-xmlfile==1.1.0
ffn==1.0.1
Flask==3.0.2
fonttools==4.50.0
frozendict==2.4.1
gcsfs
google-cloud-storage
html5lib==1.1
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.2.0
MarkupSafe==2.1.5
matplotlib==3.8.3
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pandas-datareader==0.10.0
peewee==3.17.1
pillow==10.3.0
plotly==5.20.0
pyparsing==3.1.2
PyPrind==2.11.3
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
retrying==1.3.4
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.2
xlwings==0.31.0
yfinance==0.2.37
zipp==3.18.1
